+++ src/mo_frontend/stability.ml
@@ -4,7 +4,11 @@ open Type
 
 module Pretty = Type.MakePretty(Type.ElideStampsAndHashes)
 
-let migration_link = "https://internetcomputer.org/docs/motoko/fundamentals/actors/compatibility#explicit-migration-using-a-migration-function"
+let migration_link =
+  "https://docs.internetcomputer.org/languages/motoko/fundamentals/actors/compatibility/#explicit-migration-using-a-migration-function"
+
+let enhanced_migration_link =
+  "https://docs.internetcomputer.org/languages/motoko/fundamentals/actors/enhanced-multi-migration/"
 
 (* Signature matching *)
 
@@ -24,44 +28,48 @@ let desc mig_lab_opt =
   | Some mig_lab -> "version `" ^ mig_lab ^ "`"
 
 (* FUTURE: we could perhaps use tf.src.region to better locate the errors below *)
-let error_discard s at mig_lap_opt tf =
+let error_discard s at link mig_lap_opt tf =
   Diag.add_msg s
     (Diag.error_message at "M0169" cat
-       (Format.asprintf "the stable variable `%s` of %s cannot be implicitly discarded. The variable can only be dropped by an explicit migration function, please see %s"
-        tf.lab
-        (desc mig_lap_opt)
-        migration_link))
+       (Format.asprintf
+          "stable variable `%s` of %s cannot be implicitly discarded. The variable can only be dropped by an explicit migration function.\nSee %s"
+          tf.lab
+          (desc mig_lap_opt)
+          link))
 
-let error_sub s at mig_lab_opt tf1 tf2 explanation =
+let error_sub s at link mig_lab_opt tf1 tf2 explanation =
   Diag.add_msg s
     (Diag.error_message at "M0170" cat
-      (Format.asprintf "the new type of stable variable `%s` is not compatible with %s.\n The previous type%a\n is not a subtype of%a\n because %s.\n Write an explicit migration function, please see %s."
+      (Format.asprintf
+         "stable variable `%s` is not compatible with %s.\nPrevious type%a\n is not a subtype of%a\n because %s.\nWrite an explicit migration function to convert it.\nSee %s"
          tf1.lab
         (desc mig_lab_opt)
         display_typ_expand tf1.typ
         display_typ_expand tf2.typ
         (Pretty.string_of_explanation explanation)
-        migration_link
+        link
 ))
 
-let error_stable_sub s at mig_lab_opt tf1 tf2 explanation =
+let error_stable_sub s at link mig_lab_opt tf1 tf2 explanation =
   Diag.add_msg s
     (Diag.error_message at "M0216" cat
-      (Format.asprintf "the new type of stable variable `%s` implicitly drops data of %s. \n The previous type%a\n is not a stable subtype of%a\n because %s.\n The data can only be dropped by an explicit migration function, please see %s."
+      (Format.asprintf
+         "stable variable `%s` implicitly drops data of %s.\nPrevious type%a\n is not a stable subtype of%a\n because %s.\nThe data can only be dropped by an explicit migration function.\nSee %s"
          tf1.lab
         (desc mig_lab_opt)
         display_typ_expand tf1.typ
         display_typ_expand tf2.typ
         (Pretty.string_of_explanation explanation)
-        migration_link))
+        link))
 
-let error_required s at mig_lab_opt tf =
+let error_required s at link mig_lab_opt tf =
   Diag.add_msg s
     (Diag.error_message at "M0263" cat
-       (Format.asprintf "%s does not contain the stable variable `%s`. The migration function cannot require this variable as input, please see %s."
-        (desc mig_lab_opt)
-        tf.lab
-        migration_link))
+       (Format.asprintf
+          "%s does not contain the stable variable `%s`. The migration function cannot require this variable as input.\nSee %s"
+          (desc mig_lab_opt)
+          tf.lab
+          link))
 
 (*
    - Mutability of stable fields can be changed because they are never aliased.
@@ -69,33 +77,34 @@ let error_required s at mig_lab_opt tf =
    - Lossy promotion to any or dropping record fields is rejected (stricter than subtyping to prevent data loss).
  *)
 
-let match_stab_fields s at mig_lab_opt tfs1 tfs2 =
+let match_stab_fields s at link mig_lab_opt tfs1 tfs2 =
   (* Assume that tfs1 and tfs2 are sorted. *)
   let cmp tf1 (_, tf2) = compare_field tf1 tf2 in
   Lib.List.align cmp tfs1 tfs2
     |> Seq.iter (function
       (* no dropped fields *)
       | Lib.This tf1 ->
-        error_discard s at mig_lab_opt tf1
+        error_discard s at link mig_lab_opt tf1
       (* new field ok *)
       | Lib.That (required, tf) ->
-        if required then error_required s at mig_lab_opt tf
+        if required then error_required s at link mig_lab_opt tf
       | Lib.Both (tf1, (_, tf2)) ->
         let context = [StableVariable tf2.lab] in
         begin
           match Type.sub_explained context (as_immut tf1.typ) (as_immut tf2.typ) with
-          | Incompatible explanation -> error_sub s at mig_lab_opt tf1 tf2 explanation
+          | Incompatible explanation -> error_sub s at link mig_lab_opt tf1 tf2 explanation
           | Compatible ->
              match Type.stable_sub_explained context (as_immut tf1.typ) (as_immut tf2.typ) with
-             | Incompatible explanation -> error_stable_sub s at mig_lab_opt tf1 tf2 explanation
+             | Incompatible explanation -> error_stable_sub s at link mig_lab_opt tf1 tf2 explanation
              | Compatible -> ()
         end)
 
 let incompat_mix_migrations s at =
   Diag.add_msg s
     (Diag.error_message at "M0255" cat
-        (Format.asprintf "cannot upgrade from an actor using enhanced migration to an actor not using enhanced migration. Please see %s."
-        migration_link))
+        (Format.asprintf
+           "cannot upgrade from an actor using enhanced migration to an actor not using enhanced migration.\nSee %s"
+           enhanced_migration_link))
 
 let match_stab_sig sig1 sig2 : unit Diag.result =
   match (sig1, sig2) with
@@ -109,9 +118,13 @@ let match_stab_sig sig1 sig2 : unit Diag.result =
   | _ ->
     let tfs1, mig_lab_opt = post sig1 in
     let tfs2 = pre mig_lab_opt sig2 in
+    let link = match sig2 with
+      | Multi _ -> enhanced_migration_link
+      | Single _ | PrePost _ -> migration_link
+    in
     (* Assume that tfs1 and tfs2 are sorted. *)
     let res = Diag.with_message_store (fun s ->
-      Some (match_stab_fields s no_region None tfs1 tfs2))
+      Some (match_stab_fields s no_region link None tfs1 tfs2))
     in
     (* cross check with simpler definition *)
     match res with